+++ app/views/encounter.py
