+++ tests/views/test_user.py
@@ -3,7 +3,7 @@ import unittest
 
 from app.models.item import Item
 from app.models.user import User
-from tests.controllers.controller_test_base import ControllerTestBase
+from tests.views.test_view_base import ControllerTestBase
 
 
 class UserControllerTestCase(ControllerTestBase):
